+++ CHANGELOG.md
@@ -167,3 +167,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fetch human readable bridgehead for emails
 - Email context project bridgeheads variable
 - Decode Base64 values of email context
+- Enable minus in variables of email context